+++ projects/js-packages/charts/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.1] - 2025-03-18
+### Added
+- Introduce Theme seriesLineStyles property [#42530]
+
+### Changed
+- Update package dependencies. [#42509]
+- Update package dependencies. [#42511]
+
+### Fixed
+- Bar chart: fix tickFormat not working issue [#42524]
+
 ## [0.10.0] - 2025-03-17
 ### Changed
 - Bar chart: align options with line chart [#42448]
@@ -138,6 +149,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed lints following ESLint rule changes for TS [#40584]
 - Fixing a bug in Chart storybook data. [#40640]
 
+[0.10.1]: https://github.com/Automattic/charts/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/Automattic/charts/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/Automattic/charts/compare/v0.8.4...v0.9.0
 [0.8.4]: https://github.com/Automattic/charts/compare/v0.8.3...v0.8.4