+++ projects/packages/connection/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.7.6] - 2025-03-18
+### Changed
+- Update package dependencies. [#42511]
+
 ## [6.7.5] - 2025-03-17
 ### Changed
 - Internal updates.
@@ -1354,6 +1358,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Separate the connection library into its own package.
 
+[6.7.6]: https://github.com/Automattic/jetpack-connection/compare/v6.7.5...v6.7.6
 [6.7.5]: https://github.com/Automattic/jetpack-connection/compare/v6.7.4...v6.7.5
 [6.7.4]: https://github.com/Automattic/jetpack-connection/compare/v6.7.3...v6.7.4
 [6.7.3]: https://github.com/Automattic/jetpack-connection/compare/v6.7.2...v6.7.3